Made changes as per review requests
